+++ R/mirai.R
@@ -29,16 +29,14 @@
 #' @param ... (optional) named arguments (name = value pairs) specifying
 #'     objects referenced in '.expr'. Used in addition to, and taking precedence
 #'     over, any arguments specified via '.args'.
-#' @param .args (optional) \strong{either} a list of objects to be passed by
+#' @param .args (optional) \strong{either} a list of objects passed by
 #'     \link{name} (found in the current scope), \strong{or else} a list of
-#'     name = value pairs, as in '...'. If an object other than a list is
-#'     supplied, it will be coerced to a list.
+#'     name = value pairs, as in '...'.
 #' @param .timeout [default NULL] for no timeout, or an integer value in
 #'     milliseconds. A mirai will resolve to an 'errorValue' 5 (timed out) if
 #'     evaluation exceeds this limit.
 #' @param .compute [default 'default'] character value for the compute profile
-#'     to use (each compute profile has its own set of daemons for connecting to
-#'     different resources).
+#'     to use (each compute profile has its own independent set of daemons).
 #'
 #' @return A 'mirai' object.
 #'
@@ -139,26 +137,18 @@ mirai <- function(.expr, ..., .args = list(), .timeout = NULL, .compute = "defau
   missing(.expr) && stop(._[["missing_expression"]])
 
   expr <- substitute(.expr)
-  arglist <- list(..., .expr = if (is.symbol(expr) && is.language(get0(as.character(expr), envir = sys.frame(-1L)))) .expr else expr)
-
-  if (length(.args)) {
-    if (!is.list(.args)) .args <- as.list(.args)
-    arglist <- if (length(names(.args))) c(.args, arglist) else c(`names<-`(.args, as.character(substitute(.args)[-1L])), arglist)
-  }
-
+  arglist <- list(..., .expr = if (is.symbol(expr) && is.language(.expr)) .expr else expr)
+  if (length(.args))
+    arglist <- c(if (is.null(names(.args))) `names<-`(.args, as.character(substitute(.args)[-1L])) else .args, arglist)
   data <- list2env(arglist, envir = NULL, parent = .GlobalEnv)
 
   envir <- ..[[.compute]]
-  if (length(envir)) {
-    aio <- request_signal(.context(envir[["sock"]]), data = data, cv = envir[["cv"]], send_mode = 3L, recv_mode = 1L, timeout = .timeout)
-
-  } else {
-    url <- local_url()
-    sock <- req_socket(url, resend = 0L)
-    launch_daemon(url)
+  if (is.null(envir)) {
+    sock <- ephemeral_daemon(local_url())
     aio <- request(.context(sock), data = data, send_mode = 1L, recv_mode = 1L, timeout = .timeout)
     `attr<-`(.subset2(aio, "aio"), "sock", sock)
-
+  } else {
+    aio <- request_signal(.context(envir[["sock"]]), data = data, cv = envir[["cv"]], send_mode = 3L, recv_mode = 1L, timeout = .timeout)
   }
 
   `class<-`(aio, c("mirai", "recvAio"))
@@ -205,15 +195,21 @@ everywhere <- function(.expr, ..., .args = list(), .compute = "default") {
   envir <- ..[[.compute]]
 
   if (length(envir)) {
-    expr <- c(as.expression(substitute(.expr)), .snapshot)
-    if (length(envir[["sockc"]])) {
-      expr <- c(expr, .timedelay)
-      for (i in seq_len(envir[["n"]]))
-        mirai(.expr = expr, ..., .args = .args, .compute = .compute)
-    } else {
+
+    expr <- substitute(.expr)
+    .expr <- c(as.expression(if (is.symbol(expr) && is.language(.expr)) .expr else expr), .snapshot)
+    if (length(.args) && is.null(names(.args)))
+      names(.args) <- as.character(substitute(.args)[-1L])
+
+    if (is.null(envir[["sockc"]])) {
       for (i in seq_len(max(stat(envir[["sock"]], "pipes"), envir[["n"]])))
-        mirai(.expr = expr, ..., .args = .args, .compute = .compute)
+        mirai(.expr = .expr, ..., .args = .args, .compute = .compute)
+    } else {
+      .expr <- c(.expr, .timedelay)
+      for (i in seq_len(envir[["n"]]))
+        mirai(.expr = .expr, ..., .args = .args, .compute = .compute)
     }
+
   }
 
 }
@@ -225,7 +221,8 @@ everywhere <- function(.expr, ..., .args = list(), .compute = "default") {
 #'
 #' @param aio a 'mirai' object.
 #'
-#' @return The passed mirai (invisibly). The retrieved value is stored at \code{$data}.
+#' @return The passed mirai (invisibly). The retrieved value is stored at
+#'     \code{$data}.
 #'
 #' @details This function will wait for the async operation to complete if still
 #'     in progress (blocking).
@@ -375,9 +372,9 @@ is_mirai <- function(x) inherits(x, "mirai")
 #'
 #' @return Logical value TRUE or FALSE.
 #'
-#' @details Is the object a 'miraiError'. When execution in a mirai process fails,
-#'     the error message is returned as a character string of class 'miraiError'
-#'     and 'errorValue'.
+#' @details Is the object a 'miraiError'. When execution in a mirai process
+#'     fails, the error message is returned as a character string of class
+#'     'miraiError' and 'errorValue'.
 #'
 #'     Is the object a 'miraiInterrupt'. When an ongoing mirai is sent a user
 #'     interrupt, the mirai will resolve to an empty character string classed as
@@ -447,6 +444,12 @@ print.miraiInterrupt <- function(x, ...) {
 }
 
 # internals --------------------------------------------------------------------
+
+ephemeral_daemon <- function(url) {
+  sock <- req_socket(url, resend = 0L)
+  system2(command = .command, args = c("-e", shQuote(sprintf("mirai::.daemon('%s')", url))), stdout = FALSE, stderr = FALSE, wait = FALSE)
+  sock
+}
 
 mk_interrupt_error <- function(e) .interrupt_error