+++ NEWS.md
@@ -1,3 +1,12 @@
+# mirai 0.12.1
+
+* Dispatcher initial sync timeout widened to 10s to allow for launching large numbers of daemons.
+* Default for `ssh_config()` argument 'timeout' widened to 10 (seconds).
+* Fixes `daemons()` specifying 'output = FALSE' registering as TRUE instead.
+* Fixes use of `everywhere()` specifying '.args' as an unnamed list or '.expr' as a language object.
+* Ensures compatibility with nanonext >= 0.13.0.
+* Internal performance enhancements.
+
 # mirai 0.12.0
 
 * More minimal print methods for 'mirai' and 'miraiCluster'.